src/pyproject.toml
[tool.poetry]
name = "skyone-shuge"
version = "1.0.0"
description = "天一阁 - 智能个人数字文献管理平台"
authors = ["AI Assistant <assistant@skyoneshuge.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

# Web 框架
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}

# 数据库
sqlalchemy = "^2.0.25"
aiosqlite = "^0.19.0"
psycopg2-binary = "^2.9.9"

# AI
openai = "^1.10.0"
anthropic = "^0.18.0"
langchain = "^0.1.5"

# 工具
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"

# 文档处理
pymupdf = "^1.23.0"
python-docx = "^1.1.0"

# 测试
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.8"
mypy = "^1.8.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
